# docs/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "StudyLife.Server | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/sessions": {
      "get": {
        "tags": [
          "Sessions"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StudySessionDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StudySessionDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StudySessionDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Sessions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudySessionDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StudySessionDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StudySessionDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StudySessionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudySessionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudySessionDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{id}": {
      "put": {
        "tags": [
          "Sessions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudySessionDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StudySessionDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StudySessionDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Sessions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/sessions/series/{groupId}": {
      "delete": {
        "tags": [
          "Sessions"
        ],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fromDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/sessions/history": {
      "get": {
        "tags": [
          "Sessions"
        ],
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 365
            }
          },
          {
            "name": "onlyCompleted",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StudySessionDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StudySessionDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StudySessionDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sessions/import-ics": {
      "post": {
        "tags": [
          "Sessions"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "$ref": "#/components/schemas/IFormFile"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/IcsImportResultDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IcsImportResultDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/IcsImportResultDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "IcsImportEventDto": {
        "required": [
          "title",
          "startTime",
          "endTime",
          "hasUnexpandedRecurrence"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "hasUnexpandedRecurrence": {
            "type": "boolean"
          }
        }
      },
      "IcsImportResultDto": {
        "required": [
          "events"
        ],
        "type": "object",
        "properties": {
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IcsImportEventDto"
            }
          }
        }
      },
      "IFormFile": {
        "type": "string",
        "format": "binary"
      },
...
      "StudySessionDto": {
        "required": [
          "id",
          "courseId",
          "courseName",
          "courseColor",
          "startTime",
          "endTime",
          "isCompleted",
          "timerModeId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "courseId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "courseName": {
            "type": "string"
          },
          "courseColor": {
            "type": "string"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "topic": {
            "type": [
              "null",
              "string"
            ]
          },
          "notes": {
            "type": [
              "null",
              "string"
            ]
          },
          "isCompleted": {
            "type": "boolean"
          },
          "timerModeId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "recurrenceGroupId": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "SessionToken": {
        "type": "apiKey",
        "description": "Passkey session token issued by POST /api/auth/passkey/complete or the device-link/handoff flows - what the browser/native app clients use.",
        "name": "X-Session-Token",
        "in": "header"
      },
      "ApiKey": {
        "type": "apiKey",
        "description": "Long-lived per-user API key (generated on the Setup page, one slot each for Home Assistant/AI/MCP/Capture - see ApiKeyScopes). Scoped per slot to a subset of endpoints; an out-of-scope request answers 403, not 401.",
        "name": "X-Api-Key",
        "in": "header"
      }
    }
  },
...
}